backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
supabase==2.4.2
pydantic==2.7.1
pydantic-settings==2.2.1
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dateutil==2.9.0
httpx==0.27.0
